+++ discopy/cmap.py
@@ -1347,7 +1347,7 @@ class CMap[C0: Pregroup, C1: CMap](
 
     def to_compact(self):
         """ Apply the host category's compactification functor. """
-        return self.category.compact_factory()(self)
+        return self.category.compactify_factory()(self)
 
     def to_diagram(self) -> Diagram:
         """